Move `make_run` to fixtures
