+++ autosklearn/ensemble_building/__init__.py
@@ -1,4 +1,5 @@
 from autosklearn.ensemble_building.builder import EnsembleBuilder
 from autosklearn.ensemble_building.manager import EnsembleBuilderManager
+from autosklearn.ensemble_building.run import Run
 
-__all__ = ["EnsembleBuilder", "EnsembleBuilderManager"]
+__all__ = ["EnsembleBuilder", "EnsembleBuilderManager", "Run"]
